Update from Original - 08/15/2016

+++ tools/eslint/lib/config/config-initializer.js
@@ -9,7 +9,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-var util = require("util"),
+let util = require("util"),
     debug = require("debug"),
     lodash = require("lodash"),
     inquirer = require("inquirer"),
@@ -38,14 +38,13 @@ debug = debug("eslint:config-initializer");
 function writeFile(config, format) {
 
     // default is .js
-    var extname = ".js";
+    let extname = ".js";
 
     if (format === "YAML") {
         extname = ".yml";
     } else if (format === "JSON") {
         extname = ".json";
     }
-
 
     ConfigFile.write(config, "./.eslintrc" + extname);
     log.info("Successfully created .eslintrc" + extname + " file in " + process.cwd());
@@ -61,7 +60,7 @@ function writeFile(config, format) {
  * @returns {void}
  */
 function installModules(config) {
-    var modules = [],
+    let modules = [],
         installStatus,
         modulesToInstall;
 
@@ -87,7 +86,7 @@ function installModules(config) {
 
     // Install packages which aren't already installed
     modulesToInstall = Object.keys(installStatus).filter(function(module) {
-        var notInstalled = installStatus[module] === false;
+        let notInstalled = installStatus[module] === false;
 
         if (module === "eslint" && notInstalled) {
             log.info("Local ESLint installation not found.");
@@ -114,10 +113,10 @@ function installModules(config) {
  * @returns {Object}          config object with configured rules
  */
 function configureRules(answers, config) {
-    var BAR_TOTAL = 20,
+    let BAR_TOTAL = 20,
         BAR_SOURCE_CODE_TOTAL = 4;
 
-    var newConfig = lodash.assign({}, config),
+    let newConfig = lodash.assign({}, config),
         bar,
         patterns,
         sourceCodes,
@@ -164,7 +163,7 @@ function configureRules(answers, config) {
     debug("\nRegistry: " + util.inspect(registry.rules, {depth: null}));
 
     // Create a list of recommended rules, because we don't want to disable them
-    var recRules = Object.keys(recConfig.rules).filter(function(ruleId) {
+    let recRules = Object.keys(recConfig.rules).filter(function(ruleId) {
         return ConfigOps.isErrorSeverity(recConfig.rules[ruleId]);
     });
 
@@ -203,12 +202,12 @@ function configureRules(answers, config) {
     bar.update(BAR_TOTAL);
 
     // Log out some stats to let the user know what happened
-    var finalRuleIds = Object.keys(newConfig.rules),
+    let finalRuleIds = Object.keys(newConfig.rules),
         totalRules = finalRuleIds.length;
-    var enabledRules = finalRuleIds.filter(function(ruleId) {
+    let enabledRules = finalRuleIds.filter(function(ruleId) {
         return (newConfig.rules[ruleId] !== 0);
     }).length;
-    var resultMessage = [
+    let resultMessage = [
         "\nEnabled " + enabledRules + " out of " + totalRules,
         "rules based on " + fileQty,
         "file" + ((fileQty === 1) ? "." : "s.")
@@ -226,7 +225,7 @@ function configureRules(answers, config) {
  * @returns {Object} config object
  */
 function processAnswers(answers) {
-    var config = {rules: {}, env: {}};
+    let config = {rules: {}, env: {}};
 
     if (answers.es6) {
         config.env.es6 = true;
@@ -276,10 +275,10 @@ function processAnswers(answers) {
  * @returns {Object} config object
  */
 function getConfigForStyleGuide(guide) {
-    var guides = {
+    let guides = {
         google: {extends: "google"},
         airbnb: {extends: "airbnb", plugins: ["react"]},
-        standard: {extends: "standard", plugins: ["standard"]}
+        standard: {extends: "standard", plugins: ["standard", "promise"]}
     };
 
     if (!guides[guide]) {
@@ -294,11 +293,11 @@ function getConfigForStyleGuide(guide) {
 /* istanbul ignore next: no need to test inquirer*/
 /**
  * Ask use a few questions on command prompt
- * @param {function} callback callback function when file has been written
+ * @param {Function} callback callback function when file has been written
  * @returns {void}
  */
 function promptUser(callback) {
-    var config;
+    let config;
 
     inquirer.prompt([
         {
@@ -318,7 +317,8 @@ function promptUser(callback) {
             message: "Which style guide do you want to follow?",
             choices: [{name: "Google", value: "google"}, {name: "AirBnB", value: "airbnb"}, {name: "Standard", value: "standard"}],
             when: function(answers) {
-                return answers.source === "guide";
+                answers.packageJsonExists = npmUtil.checkPackageJson();
+                return answers.source === "guide" && answers.packageJsonExists;
             }
         },
         {
@@ -342,13 +342,18 @@ function promptUser(callback) {
             default: "JavaScript",
             choices: ["JavaScript", "YAML", "JSON"],
             when: function(answers) {
-                return (answers.source === "guide" || answers.source === "auto");
+                return ((answers.source === "guide" && answers.packageJsonExists) || answers.source === "auto");
             }
         }
     ], function(earlyAnswers) {
 
         // early exit if you are using a style guide
         if (earlyAnswers.source === "guide") {
+            if (!earlyAnswers.packageJsonExists) {
+                log.info("A package.json is necessary to install plugins such as style guides. Run `npm init` to create a package.json file and try again.");
+                return;
+            }
+
             try {
                 config = getConfigForStyleGuide(earlyAnswers.styleguide);
                 writeFile(config, earlyAnswers.format);
@@ -414,7 +419,7 @@ function promptUser(callback) {
             // early exit if you are using automatic style generation
             if (earlyAnswers.source === "auto") {
                 try {
-                    var combinedAnswers = lodash.assign({}, earlyAnswers, secondAnswers);
+                    let combinedAnswers = lodash.assign({}, earlyAnswers, secondAnswers);
 
                     config = processAnswers(combinedAnswers);
                     installModules(config);
@@ -464,7 +469,7 @@ function promptUser(callback) {
                 }
             ], function(answers) {
                 try {
-                    var totalAnswers = lodash.assign({}, earlyAnswers, secondAnswers, answers);
+                    let totalAnswers = lodash.assign({}, earlyAnswers, secondAnswers, answers);
 
                     config = processAnswers(totalAnswers);
                     installModules(config);
@@ -483,7 +488,7 @@ function promptUser(callback) {
 // Public Interface
 //------------------------------------------------------------------------------
 
-var init = {
+let init = {
     getConfigForStyleGuide: getConfigForStyleGuide,
     processAnswers: processAnswers,
     initializeConfig: /* istanbul ignore next */ function(callback) {
